fix: remove the awful plugin system
The plugin system was a false good idea: it allowed farseer to be useful because you could save data to a database and do something with it.

However, the complexity to write your own plugin was just not interesting for the average user. Moreover, the API was super complicated to use & set-up. The Go docs are right, program initialization is so much harder with plugins.

By default now, farseer will do one thing & extremely well: stream message in a raw JSON format, easily exploitable.

Where you stream & what you do shouldn't be handled by farseer :)